Move refresh_hq_datasource() out of views.py
