+++ hq_superset/tasks.py
@@ -3,15 +3,13 @@ import os
 
 from superset.extensions import celery_app
 
-from .utils import AsyncImportHelper
+from .utils import AsyncImportHelper, refresh_hq_datasource
 
 logger = logging.getLogger(__name__)
 
 
 @celery_app.task(name='refresh_hq_datasource_task')
 def refresh_hq_datasource_task(domain, datasource_id, display_name, export_path, datasource_defn, user_id):
-    from .views import refresh_hq_datasource
-
     try:
         refresh_hq_datasource(domain, datasource_id, display_name, export_path, datasource_defn, user_id)
     except Exception:
